% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SPRUCE_env_files.R
\name{create_SPRUCE_env_files}
\alias{create_SPRUCE_env_files}
\title{Creates an vector object containing all SPRUCE WEW csv files of interest}
\usage{
create_SPRUCE_env_files()
}
\arguments{
\item{plot_numbers}{Numeric vector specifying the plot numbers.}

\item{date}{Character string specifying the date in the format "YYYYMMDD".}
}
\value{
a character vector containing SPRUCE environmental files
}
\description{
Creates an vector object containing all SPRUCE WEW csv files of interest
}
