% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check_function}
\alias{data_check_function}
\title{Data check function}
\usage{
data_check_function(renamed_merged_SPRUCE_df)
}
\arguments{
\item{renamed_merged_SPRUCE_df}{A data frame checked for missing values in its numeric columns.}
}
\value{
A message indicating the outcome of data validation described above.
}
\description{
Data check function
}
\details{
The function identifies numeric columns in the given data frame, checks for missing values
in these columns, and returns an error message in the console if any are found.
}
