% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_merge_function.R
\name{read_and_merge_csv_files}
\alias{read_and_merge_csv_files}
\title{Read and merge CSV files, which reads a list of CSV files into data frames, merges them into
a single dataframe, and cleans up the data a bit.}
\usage{
read_and_merge_csv_files(file_names)
}
\arguments{
\item{file_names}{A character vector specifying the names of the CSV files to be read and merged.}
}
\value{
A cleaned data frame after reading and merging the specified CSV files.
}
\description{
Read and merge CSV files, which reads a list of CSV files into data frames, merges them into
a single dataframe, and cleans up the data a bit.
}
\details{
The function reads each CSV file into a list of data frames, combines them into a single data frame using bind_rows,
andcleans up column names. Additionally, it removes a specific row with a missing value.
}
