% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols_function.R
\name{rename_cols_function}
\alias{rename_cols_function}
\title{Renaming columns to enhance comprehension}
\usage{
rename_cols_function(df)
}
\arguments{
\item{df}{A data frame that will have columns renamed.}

\item{col_rename_map}{A named list specifying the mapping of old column names to new column names.}
}
\value{
A data frame with columns renamed according to the specified column names.
}
\description{
Renaming columns to enhance comprehension
}
